Lammps move output parser to io module
