+++ pyiron_atomistics/lammps/base.py
@@ -4,17 +4,13 @@
 
 from __future__ import print_function, unicode_literals
 import os
-import posixpath
 
 import ast
-import h5py
 import numpy as np
 import pandas as pd
 import warnings
-from io import StringIO
-from dataclasses import dataclass, field, asdict
-from typing import List, Dict
 
+from pyiron_base import state, extract_data_from_file, deprecate
 from pyiron_atomistics.lammps.potential import (
     LammpsPotentialFile,
     PotentialAvailable,
@@ -22,7 +18,6 @@ from pyiron_atomistics.lammps.potential import (
     list_potentials,
 )
 from pyiron_atomistics.atomistics.job.atomistic import AtomisticGenericJob
-from pyiron_base import state, extract_data_from_file, deprecate
 from pyiron_atomistics.lammps.control import LammpsControl
 from pyiron_atomistics.lammps.potential import LammpsPotential
 from pyiron_atomistics.lammps.structure import (
@@ -31,7 +26,13 @@ from pyiron_atomistics.lammps.structure import (
     structure_to_lammps,
 )
 from pyiron_atomistics.lammps.units import UnitConverter, LAMMPS_UNIT_CONVERSIONS
-
+from pyiron_atomistics.lammps.lammps_io import (
+    collect_output_log,
+    collect_h5md_file,
+    collect_dump_file,
+    collect_errors,
+    remap_indices
+)
 
 __author__ = "Joerg Neugebauer, Sudarsan Surendralal, Jan Janssen"
 __copyright__ = (
@@ -43,22 +44,6 @@ __maintainer__ = "Sudarsan Surendralal"
 __email__ = "surendralal@mpie.de"
 __status__ = "production"
 __date__ = "Sep 1, 2017"
-
-
-@dataclass
-class DumpData:
-    steps: List = field(default_factory=lambda: [])
-    natoms: List = field(default_factory=lambda: [])
-    cells: List = field(default_factory=lambda: [])
-    indices: List = field(default_factory=lambda: [])
-    forces: List = field(default_factory=lambda: [])
-    mean_forces: List = field(default_factory=lambda: [])
-    velocities: List = field(default_factory=lambda: [])
-    mean_velocities: List = field(default_factory=lambda: [])
-    unwrapped_positions: List = field(default_factory=lambda: [])
-    mean_unwrapped_positions: List = field(default_factory=lambda: [])
-    positions: List = field(default_factory=lambda: [])
-    computes: Dict = field(default_factory=lambda: {})
 
 
 class LammpsBase(AtomisticGenericJob):
@@ -514,25 +499,10 @@ class LammpsBase(AtomisticGenericJob):
 
         """
         uc = UnitConverter(self.units)
-        prism = UnfoldingPrism(self.structure.cell, digits=15)
-        if _check_ortho_prism(prism=prism):
-            raise RuntimeError(
-                "The Lammps output will not be mapped back to pyiron correctly."
-            )
-        file_name = self.job_file_name(file_name=file_name, cwd=cwd)
-        with h5py.File(file_name, mode="r", libver="latest", swmr=True) as h5md:
-            positions = [
-                pos_i.tolist() for pos_i in h5md["/particles/all/position/value"]
-            ]
-            steps = [
-                steps_i.tolist() for steps_i in h5md["/particles/all/position/step"]
-            ]
-            forces = [for_i.tolist() for for_i in h5md["/particles/all/force/value"]]
-            # following the explanation at: http://nongnu.org/h5md/h5md.html
-            cell = [
-                np.eye(3) * np.array(cell_i.tolist())
-                for cell_i in h5md["/particles/all/box/edges/value"]
-            ]
+        forces, positions, steps, cell = collect_h5md_file(
+            file_name=self.job_file_name(file_name=file_name, cwd=cwd),
+            prism=self._prism,
+        )
         with self.project_hdf5.open("output/generic") as h5_file:
             h5_file["forces"] = uc.convert_array_to_pyiron_units(
                 np.array(forces), "forces"
@@ -560,52 +530,11 @@ class LammpsBase(AtomisticGenericJob):
         Returns:
             numpy.ndarray: Those integers mapped onto the structure.
         """
-        lammps_symbol_order = np.array(self.input.potential.get_element_lst())
-
-        # If new Lammps indices are present for which we have no species, extend the species list
-        unique_lammps_indices = np.unique(lammps_indices)
-        if len(unique_lammps_indices) > len(np.unique(self.structure.indices)):
-            unique_lammps_indices -= 1  # Convert from Lammps start counting at 1 to python start counting at 0
-            new_lammps_symbols = lammps_symbol_order[unique_lammps_indices]
-            self.structure.set_species(
-                [self.structure.convert_element(el) for el in new_lammps_symbols]
-            )
-
-        # Create a map between the lammps indices and structure indices to preserve species
-        structure_symbol_order = np.array(
-            [el.Abbreviation for el in self.structure.species]
+        return remap_indices(
+            lammps_indices=lammps_indices,
+            potential_elements=self.input.potential.get_element_lst(),
+            structure=self.structure
         )
-        map_ = np.array(
-            [
-                int(np.argwhere(lammps_symbol_order == symbol)[0]) + 1
-                for symbol in structure_symbol_order
-            ]
-        )
-
-        structure_indices = np.array(lammps_indices)
-        for i_struct, i_lammps in enumerate(map_):
-            np.place(structure_indices, lammps_indices == i_lammps, i_struct)
-        # TODO: Vectorize this for-loop for computational efficiency
-
-        return structure_indices
-
-    def collect_errors(self, file_name, cwd=None):
-        """
-
-        Args:
-            file_name:
-            cwd:
-
-        Returns:
-
-        """
-        file_name = self.job_file_name(file_name=file_name, cwd=cwd)
-        error = extract_data_from_file(file_name, tag="ERROR", num_args=1000)
-        if len(error) > 0:
-            error = " ".join(error[0])
-            raise RuntimeError("Run time error occurred: " + str(error))
-        else:
-            return True
 
     def collect_output_log(self, file_name="log.lammps", cwd=None):
         """
@@ -618,134 +547,14 @@ class LammpsBase(AtomisticGenericJob):
         Returns:
 
         """
-        uc = UnitConverter(self.units)
-        self.collect_errors(file_name=file_name, cwd=cwd)
         file_name = self.job_file_name(file_name=file_name, cwd=cwd)
+        collect_errors(file_name=file_name)
         if os.path.exists(file_name):
-            with open(file_name, "r") as f:
-                read_thermo = False
-                thermo_lines = ""
-                dfs = []
-                for l in f:
-                    l = l.lstrip()
-
-                    if read_thermo:
-                        if l.startswith("Loop"):
-                            read_thermo = False
-                            continue
-                        thermo_lines += l
-
-                    if l.startswith("Step"):
-                        read_thermo = True
-                        thermo_lines += l
-
-                dfs.append(
-                    pd.read_csv(
-                        StringIO(thermo_lines),
-                        sep="\s+",
-                        engine="c",
-                    )
-                )
-
-            if len(dfs) == 1:
-                df = dfs[0]
-            else:
-                df = pd.concat[dfs]
-
-            h5_dict = {
-                "Step": "steps",
-                "Temp": "temperature",
-                "PotEng": "energy_pot",
-                "TotEng": "energy_tot",
-                "Volume": "volume",
-            }
-
-            for key in df.columns[df.columns.str.startswith("f_mean")]:
-                h5_dict[key] = key.replace("f_", "")
-
-            df = df.rename(index=str, columns=h5_dict)
-            pressure_dict = dict()
-            if all(
-                [
-                    x in df.columns.values
-                    for x in [
-                        "Pxx",
-                        "Pxy",
-                        "Pxz",
-                        "Pxy",
-                        "Pyy",
-                        "Pyz",
-                        "Pxz",
-                        "Pyz",
-                        "Pzz",
-                    ]
-                ]
-            ):
-                pressures = (
-                    np.stack(
-                        (
-                            df.Pxx,
-                            df.Pxy,
-                            df.Pxz,
-                            df.Pxy,
-                            df.Pyy,
-                            df.Pyz,
-                            df.Pxz,
-                            df.Pyz,
-                            df.Pzz,
-                        ),
-                        axis=-1,
-                    )
-                    .reshape(-1, 3, 3)
-                    .astype("float64")
-                )
-                # Rotate pressures from Lammps frame to pyiron frame if necessary
-                if _check_ortho_prism(prism=self._prism):
-                    rotation_matrix = self._prism.R.T
-                    pressures = rotation_matrix.T @ pressures @ rotation_matrix
-
-                df = df.drop(
-                    columns=df.columns[
-                        ((df.columns.str.len() == 3) & df.columns.str.startswith("P"))
-                    ]
-                )
-                pressure_dict["pressures"] = pressures
-            else:
-                warnings.warn(
-                    "LAMMPS warning: log.lammps does not contain the required pressure values."
-                )
-            if "mean_pressure[1]" in df.columns:
-                pressures = (
-                    np.stack(
-                        (
-                            df["mean_pressure[1]"],
-                            df["mean_pressure[4]"],
-                            df["mean_pressure[5]"],
-                            df["mean_pressure[4]"],
-                            df["mean_pressure[2]"],
-                            df["mean_pressure[6]"],
-                            df["mean_pressure[5]"],
-                            df["mean_pressure[6]"],
-                            df["mean_pressure[3]"],
-                        ),
-                        axis=-1,
-                    )
-                    .reshape(-1, 3, 3)
-                    .astype("float64")
-                )
-                if _check_ortho_prism(prism=self._prism):
-                    rotation_matrix = self._prism.R.T
-                    pressures = rotation_matrix.T @ pressures @ rotation_matrix
-                df = df.drop(
-                    columns=df.columns[
-                        (
-                            df.columns.str.startswith("mean_pressure")
-                            & df.columns.str.endswith("]")
-                        )
-                    ]
-                )
-                pressure_dict["mean_pressures"] = pressures
-            generic_keys_lst = list(h5_dict.values())
+            generic_keys_lst, pressure_dict, df = collect_output_log(
+                file_name=self.job_file_name(file_name=file_name, cwd=cwd),
+                prism=self._prism
+            )
+            uc = UnitConverter(self.units)
             with self.project_hdf5.open("output/generic") as hdf_output:
                 # This is a hack for backward comparability
                 for k, v in df.items():
@@ -1054,144 +863,16 @@ class LammpsBase(AtomisticGenericJob):
         """
         uc = UnitConverter(self.units)
         file_name = self.job_file_name(file_name=file_name, cwd=cwd)
+
         if os.path.exists(file_name):
-            prism = self._prism
-            rotation_lammps2orig = self._prism.R.T
-            with open(file_name, "r") as f:
-                dump = DumpData()
-
-                for line in f:
-                    if "ITEM: TIMESTEP" in line:
-                        dump.steps.append(int(f.readline()))
-
-                    elif "ITEM: BOX BOUNDS" in line:
-                        c1 = np.fromstring(f.readline(), dtype=float, sep=" ")
-                        c2 = np.fromstring(f.readline(), dtype=float, sep=" ")
-                        c3 = np.fromstring(f.readline(), dtype=float, sep=" ")
-                        cell = np.concatenate([c1, c2, c3])
-                        lammps_cell = to_amat(cell)
-                        unfolded_cell = prism.unfold_cell(lammps_cell)
-                        dump.cells.append(unfolded_cell)
-
-                    elif "ITEM: NUMBER OF ATOMS" in line:
-                        n = int(f.readline())
-                        dump.natoms.append(n)
-
-                    elif "ITEM: ATOMS" in line:
-                        # get column names from line
-                        columns = line.lstrip("ITEM: ATOMS").split()
-
-                        # Read line by line of snapshot into a string buffer
-                        # Than parse using pandas for speed and column acces
-                        buf = StringIO()
-                        for _ in range(n):
-                            buf.write(f.readline())
-                        buf.seek(0)
-                        df = pd.read_csv(
-                            buf,
-                            nrows=n,
-                            sep="\s+",
-                            header=None,
-                            names=columns,
-                            engine="c",
-                        )
-                        df.sort_values(by="id", ignore_index=True, inplace=True)
-                        # Coordinate transform lammps->pyiron
-                        dump.indices.append(
-                            self.remap_indices(df["type"].array.astype(int))
-                        )
-
-                        force = np.stack(
-                            [df["fx"].array, df["fy"].array, df["fz"].array], axis=1
-                        )
-                        dump.forces.append(np.matmul(force, rotation_lammps2orig))
-                        if "f_mean_forces[1]" in columns:
-                            force = np.stack(
-                                [
-                                    df["f_mean_forces[1]"].array,
-                                    df["f_mean_forces[2]"].array,
-                                    df["f_mean_forces[3]"].array,
-                                ],
-                                axis=1,
-                            )
-                            dump.mean_forces.append(
-                                np.matmul(force, rotation_lammps2orig)
-                            )
-                        if "vx" in columns and "vy" in columns and "vz" in columns:
-                            v = np.stack(
-                                [
-                                    df["vx"].array,
-                                    df["vy"].array,
-                                    df["vz"].array,
-                                ],
-                                axis=1,
-                            )
-                            dump.velocities.append(np.matmul(v, rotation_lammps2orig))
-
-                        if "f_mean_velocities[1]" in columns:
-                            v = np.stack(
-                                [
-                                    df["f_mean_velocities[1]"].array,
-                                    df["f_mean_velocities[2]"].array,
-                                    df["f_mean_velocities[3]"].array,
-                                ],
-                                axis=1,
-                            )
-                            dump.mean_velocities.append(
-                                np.matmul(v, rotation_lammps2orig)
-                            )
-
-                        if "xsu" in columns:
-                            direct_unwrapped_positions = np.stack(
-                                [
-                                    df["xsu"].array,
-                                    df["ysu"].array,
-                                    df["zsu"].array,
-                                ],
-                                axis=1,
-                            )
-                            dump.unwrapped_positions.append(
-                                np.matmul(
-                                    np.matmul(direct_unwrapped_positions, lammps_cell),
-                                    rotation_lammps2orig,
-                                )
-                            )
-
-                            direct_positions = direct_unwrapped_positions - np.floor(
-                                direct_unwrapped_positions
-                            )
-                            dump.positions.append(
-                                np.matmul(
-                                    np.matmul(direct_positions, lammps_cell),
-                                    rotation_lammps2orig,
-                                )
-                            )
-
-                        if "f_mean_positions[1]" in columns:
-                            pos = np.stack(
-                                [
-                                    df["f_mean_positions[1]"].array,
-                                    df["f_mean_positions[2]"].array,
-                                    df["f_mean_positions[3]"].array,
-                                ],
-                                axis=1,
-                            )
-                            dump.mean_unwrapped_positions.append(
-                                np.matmul(
-                                    np.matmul(pos, lammps_cell),
-                                    rotation_lammps2orig,
-                                )
-                            )
-                        for k in columns:
-                            if k.startswith("c_"):
-                                kk = k.replace("c_", "")
-                                if kk not in dump.computes.keys():
-                                    dump.computes[kk] = []
-                                dump.computes[kk].append(df[k].array)
-
+            dump_dict = collect_dump_file(
+                file_name=file_name,
+                prism=self._prism,
+                structure=self.structure,
+                potential_elements=self.input.potential.get_element_lst(),
+            )
             # Write to hdf
             with self.project_hdf5.open("output/generic") as hdf_output:
-                dump_dict = asdict(dump)
 
                 for k, v in dump_dict.pop("computes").items():
                     hdf_output[k] = uc.convert_array_to_pyiron_units(
@@ -1503,65 +1184,3 @@ class Input:
             self.potential.from_hdf(hdf5_input)
             if "bond_dict" in hdf5_input.list_nodes():
                 self.bond_dict = hdf5_input["bond_dict"]
-
-
-def to_amat(l_list):
-    """
-
-    Args:
-        l_list:
-
-    Returns:
-
-    """
-    lst = np.reshape(l_list, -1)
-    if len(lst) == 9:
-        (
-            xlo_bound,
-            xhi_bound,
-            xy,
-            ylo_bound,
-            yhi_bound,
-            xz,
-            zlo_bound,
-            zhi_bound,
-            yz,
-        ) = lst
-
-    elif len(lst) == 6:
-        xlo_bound, xhi_bound, ylo_bound, yhi_bound, zlo_bound, zhi_bound = lst
-        xy, xz, yz = 0.0, 0.0, 0.0
-    else:
-        raise ValueError("This format for amat not yet implemented: " + str(len(lst)))
-
-    # > xhi_bound - xlo_bound = xhi -xlo  + MAX(0.0, xy, xz, xy + xz) - MIN(0.0, xy, xz, xy + xz)
-    # > xhili = xhi -xlo   = xhi_bound - xlo_bound - MAX(0.0, xy, xz, xy + xz) + MIN(0.0, xy, xz, xy + xz)
-    xhilo = (
-        (xhi_bound - xlo_bound)
-        - max([0.0, xy, xz, xy + xz])
-        + min([0.0, xy, xz, xy + xz])
-    )
-
-    # > yhilo = yhi -ylo = yhi_bound -ylo_bound - MAX(0.0, yz) + MIN(0.0, yz)
-    yhilo = (yhi_bound - ylo_bound) - max([0.0, yz]) + min([0.0, yz])
-
-    # > zhi - zlo = zhi_bound- zlo_bound
-    zhilo = zhi_bound - zlo_bound
-
-    cell = [[xhilo, 0, 0], [xy, yhilo, 0], [xz, yz, zhilo]]
-    return cell
-
-
-def _check_ortho_prism(prism, rtol=0.0, atol=1e-08):
-    """
-    Check if the rotation matrix of the UnfoldingPrism object is sufficiently close to a unit matrix
-
-    Args:
-        prism (pyiron_atomistics.lammps.structure.UnfoldingPrism): UnfoldingPrism object to check
-        rtol (float): relative precision for numpy.isclose()
-        atol (float): absolute precision for numpy.isclose()
-
-    Returns:
-        boolean: True or False
-    """
-    return np.isclose(prism.R, np.eye(3), rtol=rtol, atol=atol).all()
