+++ tests/lammps/test_base.py
@@ -10,7 +10,8 @@ from io import StringIO
 from pyiron_base import state, ProjectHDFio
 from pyiron_atomistics.atomistics.structure.atoms import Atoms
 from pyiron_atomistics.lammps.lammps import Lammps
-from pyiron_atomistics.lammps.base import LammpsStructure, UnfoldingPrism, to_amat
+from pyiron_atomistics.lammps.base import LammpsStructure, UnfoldingPrism
+from pyiron_atomistics.lammps.lammps_io import to_amat
 from pyiron_atomistics.lammps.units import LAMMPS_UNIT_CONVERSIONS, UnitConverter
 import ase.units as units
 from pyiron_base._tests import TestWithCleanProject
